+++ src/index.html
@@ -132,7 +132,7 @@
 <script type="module" src="../src/scripts/UserInfo.js"></script>
 <script type="module" src="../src/scripts/Api.js"></script>
 <script type="module" src="../src/scripts/Token.js"></script>
-<script type="module" src="../src/scripts/Script.js"></script>
+
 
 
 </html>
